Adds handbook-usage content
This commit adds the handbook-usage document that aims at establishing a foundation for new and existing team members on what should go on in the handbook and what should not. The content is mostely a corbon copy of GitLab's handbook usage document since it provides a good baseline for the team to build its own handbook.